# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Test",
        "description": "API for Everything",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/attendees": {
            "get": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Get all attendees",
                "operationId": "a3f8ffb08c037ce1c2b801935fed868a",
                "responses": {
                    "200": {
                        "description": "List of all attendees",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Attendee"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Create a new attendee",
                "operationId": "8a96051e7d5ec2d1618d3c050d063141",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "event_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "event_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Attendee created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Attendee"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attendees/{id}": {
            "get": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Get an attendee by ID",
                "operationId": "1005fc2e154f9ad749b4a50d90985834",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the attendee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendee details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Attendee"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Attendee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Update an attendee by ID",
                "operationId": "9ff45c860ae07a8092120afa50be8762",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the attendee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Attendee updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Attendee"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Attendee not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Delete an attendee by ID",
                "operationId": "cd9ee3d312048f109f8ba013be297a64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the attendee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendee deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Attendee deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Attendee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Attendee": {
                "title": "Attendee",
                "required": [
                    "name",
                    "email",
                    "phone",
                    "event_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the attendee",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Name of the attendee",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email of the attendee",
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "description": "Phone number of the attendee",
                        "type": "string"
                    },
                    "event_id": {
                        "description": "ID of the associated event",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token to authorize",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}